Highlight Tiles func
Have rewritten the function so it works, bit of a mess and definitely needs rewrite again. Function will be backbone of getting tiles for which player can choose to move into or pass. also resolved issue with highlighted colour - turns out it was the aplha not being set and i had this problem with last project n just forgot lol